Move metadata-related structs and functions to `metadata.rs`
Moved metadata-related structs and functions from `session.rs`
to `metadata.rs` to have only session-related functions in `session.rs`.
